Update type of "moveDistance_" and "attackRange_" variables

+++ include/npc/npc.hpp
@@ -28,13 +28,13 @@ class NPC {
     double y_ = 0;
     std::string name_ = "";
     bool alive_ = true;
-    int moveDistance_ = 1;
-    int attackRange_ = 1;
+    double moveDistance_ = 1.0;
+    double attackRange_ = 1.0;
     
   public:
     NPC();
     NPC(NPCType type);
-    NPC(NPCType type, double x, double y, const std::string &name, int moveDistance, int attackRange);
+    NPC(NPCType type, double x, double y, const std::string &name, double moveDistance, double attackRange);
 
     NPCType getType() const;
     std::string getTypeString() const;
